server/requirements.txt
django-admin
djangorestframework
markdown
django-filter
python-dotenv
dj_database_url
psycopg2-binary
requests
wheel
grpcio
setuptools
google-cloud-texttospeech
nltk
git+https://github.com/suno-ai/bark.git
google-cloud-storage
openai
